Fix Person.Id and Account.Logo type mismatches (#226)
* Fix Person.Id deserialization when API returns string instead of int

The Basecamp API returns creator.id as a JSON string in notification
responses while other endpoints return it as a number. Go's
json.Unmarshal rejects strings into int64 fields, breaking
notifications.

Add FlexibleInt64 type that accepts both JSON numbers and strings,
following the established FlexInt/FlexibleTime pattern. Applied to
Person.properties.id via enhance-openapi-go-types.sh, regenerated
Go client, and updated all service layer conversion sites.

Non-numeric sentinel strings (e.g. "basecamp" for system-generated
entities) unmarshal to zero rather than failing the response parse.
Numeric overflow strings (e.g. "9223372036854775808") are rejected.
The number path uses json.Number to avoid float64 precision loss.

* Fix Account.logo spec: string → AccountLogo object

The API returns logo as {"url": "..."} but the spec defined it as a
plain string, causing deserialization failures. Add AccountLogo
structure to the Smithy spec, update the hand-written Go Account
struct, and regenerate all 6 SDKs.

Add account-read wire-shape tests in Go, TypeScript, Ruby, and Python.
Add Go notification test with creator.id: "basecamp" sentinel to guard
the exact crash path that FlexibleInt64 fixes. Add Python multipart
upload test.

Teach the Swift and Python generators to handle multipart/form-data
by reading x-basecamp-multipart extensions. Generated multipart
headers sanitize filename (CRLF, backslash, quote) and content type
(CRLF) to prevent header injection. Python multipart helper is
method-generic.

Add missing Python account service accessor on both sync and async
clients.

+++ go/pkg/basecamp/client_approvals.go
@@ -257,7 +257,7 @@ func clientApprovalFromGenerated(ga generated.ClientApproval) ClientApproval {
 
 	if ga.Creator.Id != 0 || ga.Creator.Name != "" {
 		a.Creator = &Person{
-			ID:           ga.Creator.Id,
+			ID:           int64(ga.Creator.Id),
 			Name:         ga.Creator.Name,
 			EmailAddress: ga.Creator.EmailAddress,
 			AvatarURL:    ga.Creator.AvatarUrl,
@@ -268,7 +268,7 @@ func clientApprovalFromGenerated(ga generated.ClientApproval) ClientApproval {
 
 	if ga.Approver.Id != 0 || ga.Approver.Name != "" {
 		a.Approver = &Person{
-			ID:           ga.Approver.Id,
+			ID:           int64(ga.Approver.Id),
 			Name:         ga.Approver.Name,
 			EmailAddress: ga.Approver.EmailAddress,
 			AvatarURL:    ga.Approver.AvatarUrl,
@@ -315,7 +315,7 @@ func clientApprovalFromGenerated(ga generated.ClientApproval) ClientApproval {
 			}
 			if gr.Creator.Id != 0 || gr.Creator.Name != "" {
 				resp.Creator = &Person{
-					ID:           gr.Creator.Id,
+					ID:           int64(gr.Creator.Id),
 					Name:         gr.Creator.Name,
 					EmailAddress: gr.Creator.EmailAddress,
 					AvatarURL:    gr.Creator.AvatarUrl,
